# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the Skyrix Delivery service.",
        "title": "Skyrix Delivery API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/examples/tasks/": {
            "get": {
                "description": "Returns example tasks in descending creation order.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example tasks"
                ],
                "summary": "List example tasks",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Maximum number of tasks",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/skyrix_internal_domain_example_dto.TaskResult"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a task with PENDING status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example tasks"
                ],
                "summary": "Create an example task",
                "parameters": [
                    {
                        "description": "Task payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skyrix_internal_domain_example_dto.CreateTaskInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/skyrix_internal_domain_example_dto.TaskResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/api/v1/examples/tasks/{task_id}": {
            "get": {
                "description": "Returns an example task by its numeric identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example tasks"
                ],
                "summary": "Get an example task",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/skyrix_internal_domain_example_dto.TaskResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/api/v1/examples/tasks/{task_id}/complete": {
            "patch": {
                "description": "Changes an example task status from PENDING to COMPLETED.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example tasks"
                ],
                "summary": "Complete an example task",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/skyrix_internal_domain_example_dto.TaskResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorPayload"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_handlers.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handlers.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.ErrorPayload": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_handlers.ErrorBody"
                }
            }
        },
        "internal_handlers.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "skyrix_internal_domain_example_dto.CreateTaskInput": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 4000
                },
                "title": {
                    "type": "string",
                    "maxLength": 160
                }
            }
        },
        "skyrix_internal_domain_example_dto.TaskResult": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "public_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}